fix: Fix CSV import not accounting for overnight time entries

+++ src/main/java/me/eeshe/tempus/dto/ImportResultDTO.java
@@ -4,6 +4,6 @@ import java.util.List;
 
 public record ImportResultDTO(
         long importedCount,
-        int skippedCount,
+        long skippedCount,
         List<SkippedImportEntryDTO> skippedEntries) {
 }
